+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Version 0.8.0-SNAPSHOT - Build 20161223
 - The Dataset can now be saved and loaded.
 - All objects that save, delete and close implement the savable interfrace.
 - The InMemoryConnector now keeps a catalog of every item in a WeakReference Map.
+- In MapDBConfiguration it is now feasible to turn on/off the asynchronous writes.
 
 Version 0.7.1-SNAPSHOT - Build 20161217
 ---------------------------------------